requirements.txt
openai==1.52.2
httpx==0.27.2
pydantic==2.9.2
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.4.0
tqdm==4.66.6
pyyaml==6.0.2
